refactor(malysis): move to proper location
