+++ src/app/(malysis)/community/malysis/[analysisId]/malware-report/page.tsx
@@ -1,10 +1,11 @@
 "use client";
+
 import MalwareAnalysisReportCard from "@/components/malysis/MalwareAnalysisReportCard";
 import { GetAnalysisReportResponse } from "@buf/safedep_api.bufbuild_es/safedep/services/malysis/v1/malysis_pb";
 import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
-import MalwareAnalysisCardLoading from "./loading";
-import MalwareAnalysisError from "./error";
+import MalwareAnalysisCardLoading from "@/components/malysis/MalwareAnalysisReportLoading";
+import MalwareAnalysisError from "@/components/malysis/MalwareAnalysisError";
 
 export default function Page() {
   const { analysisId } = useParams<{ analysisId?: string }>();
@@ -13,11 +14,10 @@ export default function Page() {
   const [report, setReport] = useState<GetAnalysisReportResponse | null>();
 
   useEffect(() => {
-    async function fetchAnalysis() {
+    async function fetchReport() {
       const response = await fetch(
         `/api/community/malysis/reports/${analysisId}`,
       );
-      console.log(response);
       if (!response.ok) {
         const errorData = await response.json();
         throw new Error(errorData.message || "Failed to fetch malware report.");
@@ -26,7 +26,7 @@ export default function Page() {
       return GetAnalysisReportResponse.fromJson(data);
     }
 
-    fetchAnalysis()
+    fetchReport()
       .then((report) => {
         // setLoading(false);
         setReport(report);
@@ -50,7 +50,7 @@ export default function Page() {
   if (error) {
     return (
       <div className="flex items-center justify-center h-dvh">
-        <MalwareAnalysisError />
+        <MalwareAnalysisError error={error.message} />
       </div>
     );
   }
