+++ src/components/malysis/MalwareAnalysisError.tsx
@@ -1,12 +1,6 @@
 "use client";
-import { AlertCircle, ArrowLeft, RefreshCcw } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardFooter,
-  CardHeader,
-} from "@/components/ui/card";
+import { AlertCircle } from "lucide-react";
+import { Card, CardContent, CardHeader } from "@/components/ui/card";
 
 interface ErrorMalwareAnalysisProps {
   error?: string;
@@ -14,10 +8,8 @@ interface ErrorMalwareAnalysisProps {
   goBack?: () => void;
 }
 
-export default function MalwareAnalysisCardError({
+export default function MalwareAnalysisError({
   error = "We couldn't analyze this package. Please try again later.",
-  reset,
-  goBack,
 }: ErrorMalwareAnalysisProps) {
   return (
     <div className="container mx-auto p-4 space-y-6 max-w-6xl min-h-[50vh] grid place-items-center">
@@ -46,28 +38,6 @@ export default function MalwareAnalysisCardError({
             </div>
           </div>
         </CardContent>
-        <CardFooter className="flex gap-4 flex-wrap">
-          {reset && (
-            <Button
-              variant="default"
-              onClick={reset}
-              className="flex items-center gap-2"
-            >
-              <RefreshCcw className="h-4 w-4" />
-              Try Again
-            </Button>
-          )}
-          {goBack && (
-            <Button
-              variant="outline"
-              onClick={goBack}
-              className="flex items-center gap-2"
-            >
-              <ArrowLeft className="h-4 w-4" />
-              Go Back
-            </Button>
-          )}
-        </CardFooter>
       </Card>
     </div>
   );
