+++ src/components/malysis/MalwareAnalysisReportCard.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/table";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
-import { EvidenceRow } from "@/components/EvidenceRow";
+import EvidenceRow from "@/components/malysis/EvidenceRow";
 
 type MalwareStatus = "safe" | "possibly-malicious" | "malicious";
 
@@ -70,7 +70,7 @@ const evidences = [
   },
 ];
 
-export default function MalwareAnalysisResult() {
+export default function MalwareAnalysisReportCard() {
   const status: MalwareStatus = "safe";
   const { color, label } = statusConfig[status];
   const confidence = "Low" as "Low" | "Medium" | "High";
