+++ src/components/malysis/MalwareAnalysisReportLoading.tsx
@@ -3,9 +3,9 @@ import { Skeleton } from "@/components/ui/skeleton";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import "./styles.css";
 
-export default function LoadingMalwareAnalysis() {
+export default function MalwareAnalysisReportLoading() {
   return (
-    <div className="container mx-auto p-4 space-y-6 max-w-6xl">
+    <div className="container mx-auto p-4 space-y-6 max-w-6xl select-none">
       <Card className="border-l-4 border-l-slate-200 dark:border-l-slate-800">
         <CardHeader>
           <div className="flex items-center justify-between">
