+++ src/components/malysis/styles.css
